User UART finished
The UART for the User rover is mostly finished.  Only things left are
things we will need to change at the end of the project.  Error handling
is built into the UART driver now.  The UART driver for the User also
has the correct message queues set up and operates as it should